fixed integration group and wrongly extended KernelTestCase
