+++ tests/Entity/EntityValidationTestTrait.php
@@ -9,14 +9,14 @@
 
 namespace App\Tests\Entity;
 
-use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Validator\ConstraintViolationInterface;
 
 /**
- * @covers \App\Entity\Timesheet
+ * Classes using this MUST extend \Symfony\Bundle\FrameworkBundle\Test\KernelTestCase
  */
-abstract class AbstractEntityTest extends KernelTestCase
+trait EntityValidationTestTrait
 {
+
     /**
      * @param object $entity
      * @param array|string $fieldNames
