+++ tests/Entity/TimesheetTest.php
@@ -16,11 +16,12 @@ use App\Entity\Tag;
 use App\Entity\Timesheet;
 use App\Entity\User;
 use Doctrine\Common\Collections\ArrayCollection;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \App\Entity\Timesheet
  */
-class TimesheetTest extends AbstractEntityTest
+class TimesheetTest extends TestCase
 {
     public function testDefaultValues()
     {
@@ -64,196 +65,6 @@ class TimesheetTest extends AbstractEntityTest
         $entity->setProject($project);
 
         return $entity;
-    }
-
-    public function testValidationNeedsActivity()
-    {
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setProject(new Project())
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'activity');
-    }
-
-    public function testValidationNeedsProject()
-    {
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity(new Activity())
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'project');
-    }
-
-    public function testValidationProjectMismatch()
-    {
-        $customer = new Customer();
-        $project = (new Project())->setName('foo')->setCustomer($customer);
-        $project2 = (new Project())->setName('bar')->setCustomer($customer);
-        $activity = (new Activity())->setName('hello-world')->setProject($project);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project2)
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'project');
-    }
-
-    public function testValidationCustomerInvisible()
-    {
-        $customer = (new Customer())->setVisible(false);
-        $project = (new Project())->setName('foo')->setCustomer($customer);
-        $activity = (new Activity())->setName('hello-world')->setProject($project);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'customer');
-    }
-
-    public function testValidationCustomerInvisibleDoesNotTriggerOnStoppedEntites()
-    {
-        $customer = (new Customer())->setVisible(false);
-        $project = (new Project())->setName('foo')->setCustomer($customer);
-        $activity = (new Activity())->setName('hello-world')->setProject($project);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-            ->setEnd(new \DateTime())
-        ;
-
-        $this->assertHasNoViolations($entity);
-    }
-
-    public function testValidationProjectInvisible()
-    {
-        $customer = new Customer();
-        $project = (new Project())->setName('foo')->setCustomer($customer)->setVisible(false);
-        $activity = (new Activity())->setName('hello-world')->setProject($project);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'project');
-    }
-
-    public function testValidationProjectInvisibleDoesNotTriggerOnStoppedEntites()
-    {
-        $customer = new Customer();
-        $project = (new Project())->setName('foo')->setCustomer($customer)->setVisible(false);
-        $activity = (new Activity())->setName('hello-world')->setProject($project);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-            ->setEnd(new \DateTime())
-        ;
-
-        $this->assertHasNoViolations($entity);
-    }
-
-    public function testValidationActivityInvisible()
-    {
-        $customer = new Customer();
-        $project = (new Project())->setName('foo')->setCustomer($customer);
-        $activity = (new Activity())->setName('hello-world')->setProject($project)->setVisible(false);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-        ;
-
-        $this->assertHasViolationForField($entity, 'activity');
-    }
-
-    public function testValidationActivityInvisibleDoesNotTriggerOnStoppedEntites()
-    {
-        $customer = new Customer();
-        $project = (new Project())->setName('foo')->setCustomer($customer);
-        $activity = (new Activity())->setName('hello-world')->setProject($project)->setVisible(false);
-
-        $entity = new Timesheet();
-        $entity
-            ->setUser(new User())
-            ->setActivity($activity)
-            ->setProject($project)
-            ->setBegin(new \DateTime())
-            ->setEnd(new \DateTime())
-        ;
-
-        $this->assertHasNoViolations($entity);
-    }
-
-    public function testValidationEndNotEarlierThanBegin()
-    {
-        $entity = $this->getEntity();
-        $begin = new \DateTime();
-        $end = clone $begin;
-        $end = $end->modify('-1 second');
-        $entity->setBegin($begin);
-        $entity->setEnd($end);
-
-        $this->assertHasViolationForField($entity, 'end');
-
-        // allow same begin and end
-        $entity = $this->getEntity();
-        $begin = new \DateTime();
-        $end = clone $begin;
-        $entity->setBegin($begin);
-        $entity->setEnd($end);
-
-        $this->assertHasViolationForField($entity, []);
-    }
-
-    public function testDurationMustBeGreatorOrEqualThanZero()
-    {
-        $entity = $this->getEntity();
-        $begin = new \DateTime();
-        $end = clone $begin;
-        $entity->setBegin($begin);
-        $entity->setEnd($end);
-        $entity->setDuration(-1);
-
-        $this->assertHasViolationForField($entity, 'duration');
-
-        // allow zero duration
-        $entity = $this->getEntity();
-        $begin = new \DateTime();
-        $end = clone $begin;
-        $entity->setBegin($begin);
-        $entity->setEnd($end);
-        $entity->setDuration(0);
-
-        $this->assertHasViolationForField($entity, []);
     }
 
     public function testTags()
